+++ login/views.py
@@ -1,23 +1,36 @@
 from django.http import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
-from .models import User
+from datetime import datetime
+from .models import User, AuthToken
 from json import loads
 from hashlib import sha256
+from random import randint
 # Details
 # Takes an id as part of the endpoint
 # Returns the full details of the user with the given id (except password)
 def details(request, id):
 	# check if the request method is GET
 	if request.method == 'GET':
-		# find the user with the given id
-		user = User.objects.filter(id=id)
-		# check if the user exists
-		if user:
-			# return the serialized user data
-			return JsonResponse({'id': user[0].id, 'username': user[0].username, 'role': user[0].role}, status=200)
+		# check for an authtoken in the request headers and whether the token exists
+		token = AuthToken.objects.filter(token=request.headers.get('authtoken', ''))
+		# check if the token exists
+		if not token:
+			# return an error if the token doesn't exist
+			return JsonResponse({'error': 'You must be logged in to view this page.'}, status=401)
+		# check if the user is logged in as an admin or the user with the given id
+		if token.user_role == 'admin' or token.user_id == id:
+			# find the user with the given id
+			user = User.objects.filter(id=id)
+			# check if the user exists
+			if user:
+				# return the serialized user data
+				return JsonResponse({'id': user[0].id, 'username': user[0].username, 'role': user[0].role}, status=200)
+			else:
+				# return an error if the user doesn't exist
+				return JsonResponse({'error': 'No user found with that id.'}, status=400)
 		else:
-			# return an error if the user doesn't exist
-			return JsonResponse({'error': 'No user found with that id.'}, status=400)
+			# return an error if the user is not logged in as an admin or the user with the given id
+			return JsonResponse({'error': 'You must be logged in as an admin or the user with the given id to view this page.'}, status=401)
 	else:
 		# return an error if the request method is not GET
 		return JsonResponse({'error': 'This endpoint only accepts GET requests.'}, status=405)
@@ -28,26 +41,19 @@ def details(request, id):
 def login(request):
 	# check if the request method is POST
 	if request.method == 'POST':
-		# check if the user is already logged in
-		if request.session.get('logged_in', False) == False:
-			# get the request body and load as json
-			data = loads(request.body)
-			# try to find a user with the given username and password
-			user = User.objects.filter(username=data['username'], password=hash(data['password']))
-			# check if the user exists
-			if user:
-				# log the user in
-				request.session['logged_in'] = True
-				request.session['user_id'] = user[0].id
-				request.session['user_role'] = user[0].role
-				# return the id and role of the user
-				return JsonResponse({'result':'success', "id":user[0].id}, status=200)
-			else:
-				# return an error if the user doesn't exist
-				return JsonResponse({'error': 'No user found with that username and password.'}, status=400)
+		# get the request body and load as json
+		data = loads(request.body)
+		# try to find a user with the given username and password
+		user = User.objects.filter(username=data['username'], password=hash(data['password']))
+		# check if the user exists
+		if user:
+			# log the user in by creating a new authtoken
+			token = AuthToken.objects.create(token=generate_token(), user_id=user[0].id, user_role=user[0].role).token
+			# return the id and role of the user
+			return JsonResponse({'result':'success', "id":user[0].id, "token":token}, status=200)
 		else:
-			# return an error if a user is already logged in
-			return JsonResponse({'error': 'A user is already logged in.'}, status=409)
+			# return an error if the user doesn't exist
+			return JsonResponse({'error': 'No user found with that username and password.'}, status=400)
 	else:
 		# return an error if the request method is not POST
 		return JsonResponse({'error': 'This endpoint only accepts POST requests.'}, status=405)
@@ -56,12 +62,19 @@ def login(request):
 # Returns a success message if the user was logged out successfully
 @csrf_exempt
 def logout(request):
-	# check if the user is logged in
-	if request.session.get('logged_in', False) == True:
-		# log the user out
-		request.session.flush()
-		# return a success message
-		return JsonResponse({'result': 'success'}, status=200)
+	# check if the header contains an authtoken
+	if 'authtoken' in request.headers:
+		# find the authtoken
+		token = AuthToken.objects.filter(token=request.headers['authtoken'])
+		# check if the authtoken exists
+		if token:
+			# delete the authtoken
+			token.delete()
+			# return a success message
+			return JsonResponse({'result': 'success'}, status=200)
+		else:
+			# return an error if the authtoken doesn't exist
+			return JsonResponse({'error': 'No user logged in with that authtoken.'}, status=400)
 	else:
 		# return an error if a user is not logged in
 		return JsonResponse({'error': 'No user logged in.'}, status=409)
@@ -98,18 +111,29 @@ def register(request):
 def edit(request, id):
 	# check if the request method is PUT
 	if request.method == 'PUT':
-		# get the request body and load as json
-		data = loads(request.body)
-		user = User.objects.filter(id=id)
-		# check if the user exists
-		if user.username == data['username'] and user.password == hash(data['old_password']):
-			# update the user
-			user.update(username=data['username'], password=hash(data['password']), role=data['role'])
-			# return the id and role of the updated user
-			return JsonResponse({"result":"success","data":{'id': user[0].id, 'role': user[0].role}}, safe=False, status=200)
+		# check for an authtoken in the request headers
+		token = AuthToken.objects.filter(token=request.headers.get('authtoken', ''))
+		# check if the authtoken exists
+		if not token:
+			# return an error if the authtoken doesn't exist
+			return JsonResponse({'error': 'You must be logged in to edit a user.'}, status=401)
+		# check if the user is logged in as an admin or the user with the given id
+		if token.user_role == 'admin' or token.user_id == id:
+			# get the request body and load as json
+			data = loads(request.body)
+			user = User.objects.filter(id=id)
+			# check if the user exists
+			if user:
+				# update the user
+				user.update(username=data['username'], password=hash(data['password']), role=data['role'])
+				# return the id and role of the updated user
+				return JsonResponse({"result":"success","data":{'id': user[0].id, 'role': user[0].role}}, safe=False, status=200)
+			else:
+				# return an error if the user doesn't exist
+				return JsonResponse({'error': 'No user found with that id.'}, status=400)
 		else:
-			# return an error if the user doesn't exist
-			return JsonResponse({'error': 'Incorrect credentials.'}, status=400)
+			# return an error if the user is not logged in as an admin or the user with the given id
+			return JsonResponse({'error': 'You must be logged in as an admin or the user with the given id to view this page.'}, status=401)
 	else:
 		# return an error if the request method is not PUT
 		return JsonResponse({'error': 'This endpoint only accepts PUT requests.'}, status=405)
@@ -121,19 +145,30 @@ def edit(request, id):
 def delete(request, id):
 	# check if the request method is DELETE
 	if request.method == 'DELETE':
-		# find the user with the given id
-		user = User.objects.filter(id=id)
-		# check if the user exists
-		if user:
-			# get the id of the user to be deleted
-			id = user[0].id
-			# delete the user
-			user.delete()
-			# return a success message
-			return JsonResponse({'result': 'success', 'id':id}, status=200)
+		# check for an authtoken in the request headers
+		token = AuthToken.objects.filter(token=request.headers.get('authtoken', ''))
+		# check if the authtoken exists
+		if not token:
+			# return an error if the authtoken doesn't exist
+			return JsonResponse({'error': 'You must be logged in to delete a user.'}, status=401)
+		# check if the user is logged in as an admin or the user with the given id
+		if token.user_role == 'admin' or token.user_id == id:
+			# find the user with the given id
+			user = User.objects.filter(id=id)
+			# check if the user exists
+			if user:
+				# get the id of the user to be deleted
+				id = user[0].id
+				# delete the user
+				user.delete()
+				# return a success message
+				return JsonResponse({'result': 'success', 'id':id}, status=200)
+			else:
+				# return an error if the user doesn't exist
+				return JsonResponse({'error': 'No user found with that id.'}, status=400)
 		else:
-			# return an error if the user doesn't exist
-			return JsonResponse({'error': 'No user found with that id.'}, status=400)
+			# return an error if the user is not logged in as an admin or the user with the given id
+			return JsonResponse({'error': 'You must be logged in as an admin or the user with the given id to view this page.'}, status=401)
 	else:
 		# return an error if the request method is not DELETE
 		return JsonResponse({'error': 'This endpoint only accepts DELETE requests.'}, status=405)
@@ -144,3 +179,8 @@ def delete(request, id):
 def hash(password):
 	# hash the password
 	return sha256(password.encode('utf-8')).hexdigest()
+# Generate Token
+# Generate a token for a user
+def generate_token():
+	# generate a token
+	return sha256((str(randint(99,1000000)) + str(datetime.now)).encode('utf-8')).hexdigest()
